LayoutTests/transforms/2d/transform-2d-expected.txt

transform "none" expected "none" : PASS
transform "" expected "matrix(5.96046e-08, 1, -1, 5.96046e-08, 0, 0)" : PASS
transform "inherit" expected "none" : PASS
transform "translate(80px, 90px)" expected "matrix(1, 0, 0, 1, 80, 90)" : PASS
transform "translate(100%, 3em)" expected "matrix(1, 0, 0, 1, 200, 48)" : PASS
transform "translate(50px)" expected "matrix(1, 0, 0, 1, 50, 0)" : PASS
transform "translatex(-20px)" expected "matrix(1, 0, 0, 1, -20, 0)" : PASS
transform "translateX(-20px)" expected "matrix(1, 0, 0, 1, -20, 0)" : PASS
transform "translatey(23px)" expected "matrix(1, 0, 0, 1, 0, 23)" : PASS
transform "translateY(-3em)" expected "matrix(1, 0, 0, 1, 0, -48)" : PASS
transform "scale(1.2)" expected "matrix(1.2, 0, 0, 1.2, 0, 0)" : PASS
transform "scalex(1.5)" expected "matrix(1.5, 0, 0, 1, 0, 0)" : PASS
transform "scaleX(1.5)" expected "matrix(1.5, 0, 0, 1, 0, 0)" : PASS
transform "scaley(1.5)" expected "matrix(1, 0, 0, 1.5, 0, 0)" : PASS
transform "scaleY(1.5)" expected "matrix(1, 0, 0, 1.5, 0, 0)" : PASS
transform "scale(1.2, 0.8)" expected "matrix(1.2, 0, 0, 0.8, 0, 0)" : PASS
transform "scale(-1.2, -0.8)" expected "matrix(-1.2, -0, -0, -0.8, 0, 0)" : PASS
transform "skew(-0.7rad, 20deg)" expected "matrix(1, 0.36397, -0.842288, 1, 0, 0)" : PASS
transform "skew(12grad)" expected "matrix(1, 0, 0.19076, 1, 0, 0)" : PASS
transform "skewx(12grad)" expected "matrix(1, 0, 0.19076, 1, 0, 0)" : PASS
transform "skewX(12grad)" expected "matrix(1, 0, 0.19076, 1, 0, 0)" : PASS
transform "skewy(-12grad)" expected "matrix(1, -0.19076, 0, 1, 0, 0)" : PASS
transform "skewY(-12grad)" expected "matrix(1, -0.19076, 0, 1, 0, 0)" : PASS
transform "skewx(0.1turn)" expected "matrix(1, 0, 0.726543, 1, 0, 0)" : PASS
transform "rotate(45deg)" expected "matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0)" : PASS
transform "rotate(90deg)" expected "matrix(-4.37114e-08, 1, -1, -4.37114e-08, 0, 0)" : PASS
transform "rotate(-90deg)" expected "matrix(-4.37114e-08, -1, 1, -4.37114e-08, 0, 0)" : PASS
transform "rotate(180deg)" expected "matrix(-1, -8.74228e-08, 8.74228e-08, -1, 0, 0)" : PASS
transform "rotate(1.2rad)" expected "matrix(0.362358, 0.932039, -0.932039, 0.362358, 0, 0)" : PASS
transform "rotate(0.25turn)" expected "matrix(-4.37114e-08, 1, -1, -4.37114e-08, 0, 0)" : PASS
transform "rotate(0.5turn)" expected "matrix(-1, -8.74228e-08, 8.74228e-08, -1, 0, 0)" : PASS
transform "rotate(1.5turn)" expected "matrix(-1, -8.74228e-08, 8.74228e-08, -1, 0, 0)" : PASS
transform "matrix(1, 0, 0, 1, 0, 0)" expected "matrix(1, 0, 0, 1, 0, 0)" : PASS
transform "matrix(1, -0.19076, 0, 1, 0, 0)" expected "matrix(1, -0.19076, 0, 1, 0, 0)" : PASS
transform "matrix(0.2, 0.3, 0.4, 1.1, 1.2, 1.3)" expected "matrix(0.2, 0.3, 0.4, 1.1, 1.2, 1.3)" : PASS
transform ",rotate(12deg)" expected "none" : PASS
transform "rotate(12deg)," expected "none" : PASS
transform "rotate(12deg) +" expected "none" : PASS
transform "(translate(50, 20))" expected "none" : PASS
transform "eggs(10)" expected "none" : PASS
transform "translate(50, 20)" expected "none" : PASS
transform "translate()" expected "none" : PASS
transform "translate(10px, 20px, 30px)" expected "none" : PASS
transform "translate(10px, 20px, 30px, 40px)" expected "none" : PASS
transform "translate(10smidgens)" expected "none" : PASS
transform "translate(10px, 10smidgens)" expected "none" : PASS
transform "translateX(10px, 20px)" expected "none" : PASS
transform "translateX(10px, 20px, 30px)" expected "none" : PASS
transform "translateX(10smidgens)" expected "none" : PASS
transform "translateX(10px, 10smidgens)" expected "none" : PASS
transform "translateY(10px, 20px)" expected "none" : PASS
transform "translateY(10px, 20px, 30px)" expected "none" : PASS
transform "translateY(10smidgens)" expected "none" : PASS
transform "translateY(10px, 10smidgens)" expected "none" : PASS
transform "scale(2px, 2px)" expected "none" : PASS
transform "scale(2, 2px)" expected "none" : PASS
transform "scale()" expected "none" : PASS
transform "scale(1, 2, 3)" expected "none" : PASS
transform "scale(2smidgens)" expected "none" : PASS
transform "scale(2, 2smidgens)" expected "none" : PASS
transform "rotate(10)" expected "none" : PASS
transform "rotate()" expected "none" : PASS
transform "rotate(10deg, 20deg)" expected "none" : PASS
transform "rotate(2turns)" expected "none" : PASS
transform "rotate(2spins)" expected "none" : PASS
transform "rotate(2, 2spins)" expected "none" : PASS
transform "skew(10)" expected "none" : PASS
transform "skew()" expected "none" : PASS
transform "skew(10deg, 20deg, 30deg)" expected "none" : PASS
transform "skew(2bits)" expected "none" : PASS
transform "skew(2, 2bits)" expected "none" : PASS
transform "skewX(10)" expected "none" : PASS
transform "skewX()" expected "none" : PASS
transform "skewX(10deg, 20deg)" expected "none" : PASS
transform "skewX(2bits)" expected "none" : PASS
transform "skewY(10)" expected "none" : PASS
transform "skewY()" expected "none" : PASS
transform "skewY(10deg, 20deg)" expected "none" : PASS
transform "skewY(2bits)" expected "none" : PASS
transform "matrix()" expected "none" : PASS
transform "matrix(2, 0, 0, 2)" expected "none" : PASS
transform "matrix(0.978148, 0.207912, -0.207912, 0.978148, 50, 20, 666)" expected "none" : PASS
transform "matrix(1, 0, 0, 1, 20px, 50px)" expected "none" : PASS
